Move shared logic to an abstract base class
