+++ src/java/net/jpountz/lz4/LZ4JNIHCFastResetCompressor.java
@@ -19,11 +19,7 @@ package net.jpountz.lz4;
 import static net.jpountz.lz4.LZ4Constants.DEFAULT_COMPRESSION_LEVEL;
 import static net.jpountz.lz4.LZ4Constants.MAX_COMPRESSION_LEVEL;
 
-import net.jpountz.util.ByteBufferUtils;
-import net.jpountz.util.SafeUtils;
-
 import java.nio.ByteBuffer;
-import java.util.concurrent.locks.ReentrantLock;
 
 /**
  * An optimized LZ4 HC compressor that uses native {@code LZ4_compress_HC_extStateHC_fastReset}.
@@ -57,14 +53,8 @@ import java.util.concurrent.locks.ReentrantLock;
  * @see LZ4Factory#highFastResetCompressor(int)
  * @see LZ4HCJNICompressor
  */
-public final class LZ4JNIHCFastResetCompressor extends LZ4Compressor implements AutoCloseable {
+public final class LZ4JNIHCFastResetCompressor extends AbstractLZ4JNIFastResetCompressor {
 
-  private static final String IN_USE_ERROR = "This compressor is not thread-safe and is already in use";
-  private static final String CLOSED_ERROR = "Compressor has been closed";
-  private static final String UNSUPPORTED_BUFFER_ERROR = "ByteBuffer must be direct or array-backed";
-
-  private final ReentrantLock lock = new ReentrantLock();
-  private long statePtr;
   private final int compressionLevel;
 
   /**
@@ -82,17 +72,13 @@ public final class LZ4JNIHCFastResetCompressor extends LZ4Compressor implements 
    * @param compressionLevel compression level (1-17, higher = better compression)
    */
   LZ4JNIHCFastResetCompressor(int compressionLevel) {
+    super(LZ4JNI.LZ4_createStreamHC(), "Failed to allocate LZ4 HC state");
     if (compressionLevel > MAX_COMPRESSION_LEVEL) {
       compressionLevel = MAX_COMPRESSION_LEVEL;
     } else if (compressionLevel < 1) {
       compressionLevel = DEFAULT_COMPRESSION_LEVEL;
     }
     this.compressionLevel = compressionLevel;
-    long ptr = LZ4JNI.LZ4_createStreamHC();
-    if (ptr == 0) {
-      throw new LZ4Exception("Failed to allocate LZ4 HC state");
-    }
-    this.statePtr = ptr;
   }
 
   /**
@@ -104,144 +90,17 @@ public final class LZ4JNIHCFastResetCompressor extends LZ4Compressor implements 
     return compressionLevel;
   }
 
-  /**
-   * Compresses {@code src[srcOff:srcOff+srcLen]} into
-   * {@code dest[destOff:destOff+maxDestLen]}.
-   *
-   * @param src        source data
-   * @param srcOff     the start offset in src
-   * @param srcLen     the number of bytes to compress
-   * @param dest       destination buffer
-   * @param destOff    the start offset in dest
-   * @param maxDestLen the maximum number of bytes to write in dest
-   * @return the compressed size
-   * @throws LZ4Exception if maxDestLen is too small
-   * @throws IllegalStateException if the compressor has been closed or is already in use
-   */
-  public int compress(byte[] src, int srcOff, int srcLen, byte[] dest, int destOff, int maxDestLen) {
-    if (!lock.tryLock()) {
-      throw new IllegalStateException(IN_USE_ERROR);
-    }
-
-    try {
-      if (statePtr == 0) {
-        throw new IllegalStateException(CLOSED_ERROR);
-      }
-      SafeUtils.checkRange(src, srcOff, srcLen);
-      SafeUtils.checkRange(dest, destOff, maxDestLen);
-
-      final int result = LZ4JNI.LZ4_compress_HC_extStateHC_fastReset(
-        statePtr, src, null, srcOff, srcLen,
-        dest, null, destOff, maxDestLen, compressionLevel);
-
-      if (result <= 0) {
-        throw new LZ4Exception("maxDestLen is too small");
-      }
-      return result;
-    } finally {
-      lock.unlock();
-    }
-  }
-
-  /**
-   * Compresses {@code src[srcOff:srcOff+srcLen]} into
-   * {@code dest[destOff:destOff+maxDestLen]}.
-   * <p>
-   * Both buffers must be either direct or array-backed.
-   * {@link ByteBuffer} positions remain unchanged.
-   *
-   * @param src        source data
-   * @param srcOff     the start offset in src
-   * @param srcLen     the number of bytes to compress
-   * @param dest       destination buffer
-   * @param destOff    the start offset in dest
-   * @param maxDestLen the maximum number of bytes to write in dest
-   * @return the compressed size
-   * @throws LZ4Exception if maxDestLen is too small
-   * @throws IllegalArgumentException if src or dest is neither array-backed nor direct
-   * @throws IllegalStateException if the compressor has been closed or is already in use
-   */
-  public int compress(ByteBuffer src, int srcOff, int srcLen, ByteBuffer dest, int destOff, int maxDestLen) {
-    if (!lock.tryLock()) {
-      throw new IllegalStateException(IN_USE_ERROR);
-    }
-
-    try {
-      if (statePtr == 0) {
-        throw new IllegalStateException(CLOSED_ERROR);
-      }
-      checkByteBuffer(src);
-      checkByteBuffer(dest);
-      ByteBufferUtils.checkNotReadOnly(dest);
-      ByteBufferUtils.checkRange(src, srcOff, srcLen);
-      ByteBufferUtils.checkRange(dest, destOff, maxDestLen);
-
-      return compressNativeBuffers(statePtr, src, srcOff, srcLen, dest, destOff, maxDestLen);
-    } finally {
-      lock.unlock();
-    }
-  }
-
-  private int compressNativeBuffers(long ptr, ByteBuffer src, int srcOff, int srcLen,
-                                    ByteBuffer dest, int destOff, int maxDestLen) {
-    byte[] srcArr = src.hasArray() ? src.array() : null;
-    byte[] destArr = dest.hasArray() ? dest.array() : null;
-    ByteBuffer srcBuf = srcArr == null ? src : null;
-    ByteBuffer destBuf = destArr == null ? dest : null;
-    int srcBufferOff = srcOff + (srcArr != null ? src.arrayOffset() : 0);
-    int destBufferOff = destOff + (destArr != null ? dest.arrayOffset() : 0);
-
-    final int result = LZ4JNI.LZ4_compress_HC_extStateHC_fastReset(
-      ptr, srcArr, srcBuf, srcBufferOff, srcLen,
-      destArr, destBuf, destBufferOff, maxDestLen, compressionLevel);
-
-    if (result <= 0) {
-      throw new LZ4Exception("maxDestLen is too small");
-    }
-    return result;
-  }
-
-  private static void checkByteBuffer(ByteBuffer buffer) {
-    if (!(buffer.hasArray() || buffer.isDirect())) {
-      throw new IllegalArgumentException(UNSUPPORTED_BUFFER_ERROR);
-    }
-  }
-
-  /**
-   * Returns true if this compressor has been closed.
-   *
-   * @return true if closed
-   */
-  public boolean isClosed() {
-    lock.lock();
-    try {
-      return statePtr == 0;
-    } finally {
-      lock.unlock();
-    }
-  }
-
-  /**
-   * Closes this compressor and releases native resources.
-   * After calling this method, all compress methods will throw {@link IllegalStateException}.
-   *
-   * @throws IllegalStateException if the compressor is in use by another thread
-   */
   @Override
-  public void close() {
-    if (!lock.tryLock()) {
-      throw new IllegalStateException(IN_USE_ERROR);
-    }
+  protected int compressNative(long ptr, byte[] srcArr, ByteBuffer srcBuf, int srcOff, int srcLen,
+                               byte[] destArr, ByteBuffer destBuf, int destOff, int maxDestLen) {
+    return LZ4JNI.LZ4_compress_HC_extStateHC_fastReset(
+      ptr, srcArr, srcBuf, srcOff, srcLen,
+      destArr, destBuf, destOff, maxDestLen, compressionLevel);
+  }
 
-    try {
-      long ptr = statePtr;
-      statePtr = 0;
-      if (ptr != 0) {
-        LZ4JNI.LZ4_freeStreamHC(ptr);
-      }
-    } finally {
-      lock.unlock();
-    }
+  @Override
+  protected void freeState(long ptr) {
+    LZ4JNI.LZ4_freeStreamHC(ptr);
   }
 
   @Override
